requirements.txt
# Основные библиотеки
discord.py==2.3.2
yt-dlp==2025.4.30
PyNaCl==1.5.0
async-timeout==4.0.3
python-dotenv==1.0.0
certifi>=2023.7.22
requests>=2.29.0

# Для установки с отключенной проверкой SSL используйте:
# pip install -r requirements.txt --trusted-host pypi.org --trusted-host files.pythonhosted.org
